add a test
