+++ packages/next/src/index.ts
@@ -66,7 +66,6 @@ import {
   getFilesMapFromReasons,
   getImagesConfig,
   getImagesManifest,
-  getMiddlewareManifest,
   getNextConfig,
   getPageLambdaGroups,
   getPrerenderManifest,
@@ -91,8 +90,8 @@ import {
   getOperationType,
   isApiPage,
   getFunctionsConfigManifest,
-  normalizeEdgeFunctionPath,
   require_,
+  getServerlessPages,
 } from './utils';
 
 export const version = 2;
@@ -2683,60 +2682,3 @@ export const prepareCache: PrepareCache = async ({
   debug('Cache file manifest produced');
   return cache;
 };
-
-async function getServerlessPages(params: {
-  pagesDir: string;
-  entryPath: string;
-  outputDirectory: string;
-  appPathRoutesManifest?: Record<string, string>;
-}) {
-  const appDir = path.join(params.pagesDir, '../app');
-  const [pages, appPaths, middlewareManifest] = await Promise.all([
-    glob('**/!(_middleware).js', params.pagesDir),
-    params.appPathRoutesManifest
-      ? Promise.all([
-          glob('**/page.js', appDir),
-          glob('**/route.js', appDir),
-          glob('**/_not-found.js', appDir),
-        ]).then(items => Object.assign(...items))
-      : Promise.resolve({}),
-    getMiddlewareManifest(params.entryPath, params.outputDirectory),
-  ]);
-
-  const normalizedAppPaths: typeof appPaths = {};
-
-  if (params.appPathRoutesManifest) {
-    for (const [entry, normalizedEntry] of Object.entries(
-      params.appPathRoutesManifest
-    )) {
-      const normalizedPath = `${path.join(
-        '.',
-        normalizedEntry === '/' ? '/index' : normalizedEntry
-      )}.js`;
-      const globPath = `${path.posix.join('.', entry)}.js`;
-
-      if (appPaths[globPath]) {
-        normalizedAppPaths[normalizedPath] = appPaths[globPath];
-      }
-    }
-  }
-
-  // Edge Functions do not consider as Serverless Functions
-  for (const edgeFunctionFile of Object.keys(
-    middlewareManifest?.functions ?? {}
-  )) {
-    let edgePath =
-      middlewareManifest?.functions?.[edgeFunctionFile].name ||
-      edgeFunctionFile;
-
-    edgePath = normalizeEdgeFunctionPath(
-      edgePath,
-      params.appPathRoutesManifest || {}
-    );
-    edgePath = (edgePath || 'index') + '.js';
-    delete normalizedAppPaths[edgePath];
-    delete pages[edgePath];
-  }
-
-  return { pages, appPaths: normalizedAppPaths };
-}
