+++ packages/next/src/utils.ts
@@ -3063,3 +3063,60 @@ export async function getVariantsManifest(
 
   return variantsManifest;
 }
+
+export async function getServerlessPages(params: {
+  pagesDir: string;
+  entryPath: string;
+  outputDirectory: string;
+  appPathRoutesManifest?: Record<string, string>;
+}) {
+  const appDir = path.join(params.pagesDir, '../app');
+  const [pages, appPaths, middlewareManifest] = await Promise.all([
+    glob('**/!(_middleware).js', params.pagesDir),
+    params.appPathRoutesManifest
+      ? Promise.all([
+          glob('**/page.js', appDir),
+          glob('**/route.js', appDir),
+          glob('**/_not-found.js', appDir),
+        ]).then(items => Object.assign(...items))
+      : Promise.resolve({}),
+    getMiddlewareManifest(params.entryPath, params.outputDirectory),
+  ]);
+
+  const normalizedAppPaths: typeof appPaths = {};
+
+  if (params.appPathRoutesManifest) {
+    for (const [entry, normalizedEntry] of Object.entries(
+      params.appPathRoutesManifest
+    )) {
+      const normalizedPath = `${path.join(
+        '.',
+        normalizedEntry === '/' ? '/index' : normalizedEntry
+      )}.js`;
+      const globPath = `${path.posix.join('.', entry)}.js`;
+
+      if (appPaths[globPath]) {
+        normalizedAppPaths[normalizedPath] = appPaths[globPath];
+      }
+    }
+  }
+
+  // Edge Functions do not consider as Serverless Functions
+  for (const edgeFunctionFile of Object.keys(
+    middlewareManifest?.functions ?? {}
+  )) {
+    let edgePath =
+      middlewareManifest?.functions?.[edgeFunctionFile].name ||
+      edgeFunctionFile;
+
+    edgePath = normalizeEdgeFunctionPath(
+      edgePath,
+      params.appPathRoutesManifest || {}
+    );
+    edgePath = (edgePath || 'index') + '.js';
+    delete normalizedAppPaths[edgePath];
+    delete pages[edgePath];
+  }
+
+  return { pages, appPaths: normalizedAppPaths };
+}
