+++ packages/next/test/unit/utils.test.ts
@@ -1,12 +1,15 @@
 import path from 'path';
+import os from 'os';
 import {
   excludeFiles,
   validateEntrypoint,
   normalizePackageJson,
   getImagesConfig,
   getNextConfig,
+  getServerlessPages,
 } from '../../src/utils';
 import { FileRef } from '@vercel/build-utils';
+import { genDir } from '../utils';
 
 describe('getNextConfig', () => {
   const workPath = path.join(__dirname, 'fixtures', '00-config');
@@ -378,5 +381,30 @@ describe('normalizePackageJson', () => {
         testEnvironment: 'node',
       },
     });
+  });
+});
+
+describe('getServerlessPages', () => {
+  it('should gather all pages correctly', async () => {
+    const dir = await genDir({
+      '.next/server/pages/_app.js': 'test',
+      '.next/server/pages/_error.js': 'test',
+      '.next/server/app/page.js': 'test',
+      '.next/server/app/favicon.ico/route.js': 'test',
+    });
+
+    const { pages, appPaths } = await getServerlessPages({
+      pagesDir: path.resolve(path.join(dir, '.next/server/pages')),
+      entryPath: os.tmpdir(),
+      outputDirectory: os.tmpdir(),
+      appPathRoutesManifest: {
+        '/_not-found': '/_not-found',
+        '/favicon.ico/route': '/favicon.ico',
+        '/page': '/',
+      },
+    });
+
+    expect(Object.keys(pages)).toEqual(['_app.js', '_error.js']);
+    expect(Object.keys(appPaths)).toEqual(['favicon.ico.js', 'index.js']);
   });
 });
